Initial commit + slight refactor
