+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/actions/provider/ViewProviderActions.java
@@ -325,10 +325,6 @@ public class ViewProviderActions
                     provider.getString("owner") +
                     ") - " +
                     provider.getString("status");
-            default -> {
-                String msg = String.format("Unsupported provider type (%s).", providerType);
-                throw new IllegalStateException(msg);
-            }
         };
     }
